requirements.txt
# BFRM - Bubble Flow Reconstruction and Measurement
# Real-time Version Dependencies

# Core GUI Framework
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep Learning and YOLO
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# 3D Visualization
pyvista>=0.42.0
pyvistaqt>=0.11.0
vtk>=9.2.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0

# Plotting and Visualization
plotly>=5.15.0
seaborn>=0.12.0

# System Monitoring
psutil>=5.9.0

# File I/O and Utilities
tqdm>=4.65.0
pillow>=10.0.0
imageio>=2.31.0

# Optional: GPU Acceleration
# CUDA Toolkit should be installed separately
# cupy-cuda11x>=12.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x

# Development and Testing (Optional)
pytest>=7.4.0
pytest-qt>=4.2.0
black>=23.0.0
flake8>=6.0.0

# Documentation (Optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0 